Abstract list into separate command
